od: start supporting getting SRD from survey info provider
Summary:
SIP now has a new virtual function to get SRD.
Needs work in unit section.

Test Plan: To be tested by Arnaud

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D7209

git-svn-id: https://svn.opendtect.org/od/trunk@38344 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Basic/survinfo.h
@@ -123,10 +123,10 @@ public:
     void		snapZ(float&,int direction=0) const;
 			//!< see snap() for direction
 
-    double		seismicReferenceDatum() const	 {return seisrefdatum_;}
+    float		seismicReferenceDatum() const	 {return seisrefdatum_;}
 			/*!<In depth units (m or ft), positive upward
-			from sea level */
-    void		setSeismicReferenceDatum(double d){ seisrefdatum_=d; }
+			    from sea level */
+    void		setSeismicReferenceDatum( float d ) { seisrefdatum_=d; }
 
     const IOPar&	pars() const			{ return pars_; }
     void		putZDomain(IOPar&) const;
@@ -150,7 +150,7 @@ protected:
     bool		depthsinfeet_;
     TrcKeyZSampling&	tkzs_;
     TrcKeyZSampling&	wcs_;
-    double		seisrefdatum_;
+    float		seisrefdatum_;
     IOPar&		pars_;
 
     mutable Threads::AtomicPointer<Survey::Geometry3D>	s3dgeom_;
